Add a script to join all runs in current folder
SubtitleEdit can do this as well, but it always takes quite a few manual
UI actions. This script is just faster, and by default it generates a
new file (depending on the time) so that the previous merge can be
compared.

It also has a second trick: it has a "-debugId" flag which will write
special subtitle sequence numbers to help find the right run file to
edit after an error gets spotted in the merged file. The first digits
indicate the number of the run file, then there's an extra 0 as padding,
then there's the sequence number of the subtitle in the original run
file. This is especially handy for when very short runs overlap.