Index: Finetune offset diff by syncing a specific point in the audio stream
Comparing the start of the first subtitle with the original SRT file
might be deceiving: sometimes that first line is in the wrong spot so
the offset difference is wrong.

The other important bit is that different SRT files might have different
"lead-in" times; one language might show a line 100ms before the words
are spoken while another language's SRT might prefer to show it at the
exact moment the words are spoken. As a result, different languages
would need corrections applied.

Syncing on the actual audio stream avoids this problem.

While at it, split the index spreadsheet into two working sheets: one
contains an overview and descriptions of all the runs, the other
contains the raw data + all the math. Lookups copy the raw data + math
results to the overview sheet.