Index: rework overview page to use vlookups, use the "Fix" colum up to r093
So far the offsets were chosen arbitrarily. Now that we have data about
the video and audio syncing issues we can opt to shift subtitles to be
in sync with the audio plus or minus a small correction value.

Picking the correction value is a trade-off between keeping the
subtitles in a camera shot and actually having it start on time
according to the sound. Points where the subtitles don't need strict
aligment to the many short camera shots is when no correction should be
done, at others it's best to correct so that the subtitle is still too
late (character is already moving their lips) but at least the text is
gone when the camera shot changes.

Runs from r094 on still need their values tweaked, for now they're based
only on getting the subtitles to fit with the sound.