Scripts: Implement a script to strip name tags and captions from srt files
This is used to convert the EN-CC version into a plain English subtitle
file without having to duplicate all the different pieces. Any bugfixes
to typing mistakes are thus immediately fixed in both versions.

The conversion script needs at least one file in a lanuage directory:
config.psm1
This script module defines which other folder the subtitles are taken
from (should we ever want to reuse this for another language).

Then each srt file in the source folder will be looked up, and "command
files" in the target language folder will define what need to happen to
them (basically, the extension defines the type of command):
- rxxx.copy means the source file will be copied to the output as-is
- rxxx.del means the source file is ignored / deleted from the output
- rxxx.replace means the contents of this fill will replace the source
- rxxx.filter will remove names and captions using the default filter
- rxxx.psm1 allows a custom filter script