Index: add column to indicate runs reserved for closed captions
Earlier versions of the index calculated time shifting offsets until the
first spoken bit of dialog, but this made it hard to sync when there
were captions at the start. As a result, these runs were split off into
their own separate srt.

When adding a new translation it's useful to clearly indicate this kind
of run in a dedicated column to make it more obvious that these don't
contain any dialog.