Index: Fix data entry error for run 7 and 28
Run 7:
The "Source end" time was wrong because a fixed value was pasted where a
formula was supposed to be.

Also mark it as CC-only, it's not all that useful.

Run 28:
In retrospect this run is only really useful for closed captions.
Translations will never need to translate that short moan.