Scripts: Script to split 6 source srt files into TBE subtitle runs
This script automates what had to be done by hand earlier on: finding
the nearest time offsets on which to start copying, finding the nearest
time offset on which to stop and then saving everything in a new "run"
file.

Some data from the index file and a few clever sorting tricks allow to
extract all that info rather quickly.

Do mind that there might be subtitles which hang over the start or end
of a run. This might be a sign that the index file has a run start and
end offset which is too conservative, or the subtitle contains more text
than expected so that it needs to be split.