Index / Scripts: Tweak timings to have more leeway
Different subtitles might have different lead times and display lengths.
The old file was built on exact timings from one particular subtitle,
but this is not very useful for automatically converting translated
subtitles.

The new tweaked timings are rounded up or down so that the cutting
script triggers less warnings. A few runs end in the middle of a dialog,
and then there might still be a warning indicating that the freshly cut
translation might contain too much text.