NL E3.1: Edit till end, make credits vague again
This completes the editing step where we merge / split subtitles to fit
or really move them around.

SubtitleEdit suggested joining lines, fixing italic tags, etc.

Then there were some captions in r108 which could be adjusted so that
their CPM is better.

Also, it's not known who made the translation of the first movie which this
subtitle project is based on. It is known for the second and third
movies, but the quality of those translations seems to be lower than the
quality of the first movie.

As a result it might be better to include no names at all, that way it
cannot be misattributed.