Scripts: convert-sync-points formats sync srt files into a simple TSV file
The sync points have always been stored in subtitle files for easy
referencing. Then the timestamps of each point needed to be copy/pasted
one by one into the index spreadsheet or the SRT file was converted to a
simple list using a Vim macro.

This script automates the creation of these lists of sync points so that
the list can be opened as a TSV file using a spreadsheet program. The
list can then be easily copied as a whole and pasted into the index.

This reduces the chance of mistakes (as has happened in the past).

Also update .gitignore to ignore these new TSV files. Also ignore Vegas
Pro files because it's the tool which is currently used to create more
precise sync points.